fix(core): remove yaml observer support
